+++ app/Http/Controllers/Events/EventCalendar.php
@@ -5,15 +5,16 @@ declare(strict_types=1);
 namespace App\Http\Controllers\Events;
 
 use App\Http\Controllers\Controller;
+use Carbon\Carbon;
 
 class EventCalendar extends Controller
 {
     public function render()
     {
-        $jsonEventList = '[]';
+        $startOfLastYear = Carbon::now()->subYear()->setDay(1)->setMonth(1)->setTime(0, 0);
 
         $eventList = \App\Models\Event::getAllFiltered(
-            new \App\Models\EventFilter(true, false, ! auth()->guest())
+            new \App\Models\Filter\EventFilter($startOfLastYear, null, true, false, !auth()->guest())
         )
             ->get(['title', 'start', 'end', 'whole_day', 'description', 'location', 'dress_code'])
             ->map(function ($event) {