+++ app/Livewire/EventFilterTrait.php
@@ -3,23 +3,21 @@
 namespace App\Livewire;
 
 use App\Models\Event;
-use App\Models\EventFilter;
+use App\Models\Filter\EventFilter;
+use Carbon\Carbon;
 use Illuminate\Support\Facades\Auth;
 
 trait EventFilterTrait
 {
-    public bool $sortAsc = true;
+    public ?string $start = null;
 
-    public bool $showPast = false;
+    public ?string $end = null;
+
+    public bool $sortAsc = true;
 
     public bool $showDisabled = false;
 
     public bool $showLoggedInOnly = true;
-
-    public function canFilterShowPast(): bool
-    {
-        return true;
-    }
 
     public function canFilterShowDisabled(): bool
     {
@@ -34,7 +32,8 @@ trait EventFilterTrait
     public function getEventFilter(): EventFilter
     {
         return new EventFilter(
-            $this->canFilterShowPast() ? $this->showPast : false,
+            $this->start ? Carbon::parseFromDatetimeLocalInput($this->start) : null,
+            $this->end ? Carbon::parseFromDatetimeLocalInput($this->end) : null,
             $this->canFilterShowDisabled() ? $this->showDisabled : false,
             $this->canFilterShowLoggedInOnly() ? $this->showLoggedInOnly : false,
             $this->sortAsc